Separate BlockBasedModification into its own file
